+++ module/config.mjs
