--- module/index.js
@@ -1,30 +1,0 @@
-import * as layout from './layout/index.js';
-import * as connectors from './connectors/index.js';
-import * as highlighters from './highlighters/index.js';
-import * as connectionPoints from './connectionPoints/index.js';
-import * as connectionStrategies from './connectionStrategies/index.js';
-import * as routers from './routers/index.js';
-import * as anchors from './anchors/index.js';
-import * as linkAnchors from './linkAnchors/index.mjs';
-import * as dia from './dia/index.js';
-import * as shapes from './shapes/index.js';
-import * as linkTools from './linkTools/index.js';
-import * as util from './util/index.js';
-import * as mvc from './mvc/index.js';
-
-export * from './base.js';
-export {
-    anchors,
-    linkAnchors,
-    connectionPoints,
-    connectionStrategies,
-    connectors,
-    dia,
-    highlighters,
-    layout,
-    linkTools,
-    routers,
-    shapes,
-    util,
-    mvc
-}
