+++ wrappers/joint-dia-esm.head.js
@@ -1,20 +1,20 @@
 import $ from 'jquery';
 import * as BackboneModule from 'backbone';
 import * as g from '../../src/geometry.js';
-import * as util from './../util/index.js';
-import * as mvc from '../mvc/index.js';
+import * as util from './../util/index.mjs';
+import * as mvc from '../mvc/index.mjs';
 import V from '../../src/vectorizer.js';
 import * as joint from '../../src/core.js';
 import { ToolView, ToolsView } from '../../src/joint.dia.tools.js';
 
 import { Port } from '../layout/port.js';
 import { PortLabel } from '../layout/portLabel.js';
-import * as connectors from '../connectors/index.js';
-import * as highlighters from '../highlighters/index.js';
-import * as connectionPoints from '../connectionPoints/index.js';
-import * as connectionStrategies from '../connectionPoints/index.js';
-import * as routers from '../routers/index.js';
-import * as anchors from '../anchors/index.js';
+import * as connectors from '../connectors/index.mjs';
+import * as highlighters from '../highlighters/index.mjs';
+import * as connectionPoints from '../connectionpoints/index.mjs';
+import * as connectionStrategies from '../connectionpoints/index.mjs';
+import * as routers from '../routers/index.mjs';
+import * as anchors from '../anchors/index.mjs';
 
 const Backbone = Object.assign({}, BackboneModule.default);
